refactor: update decorators dir name and add user interface
