+++ backend/src/common/interfaces/user.interface.ts
@@ -1,0 +1,7 @@
+export interface UserPayload {
+  user: {
+    sub: string;
+    email: string;
+    role: string;
+  };
+}
